modify

+++ src/ch/min2phase/Search.java
@@ -47,7 +47,7 @@ public class Search {
     private boolean inited = false;
 
     // 值范围: [0, 17]
-    private int[] move = new int[31];
+    private byte[] move = new byte[31];
 
     private CoordCube[] nodeUD = new CoordCube[MAX_PRE_MOVES];
 
@@ -92,7 +92,7 @@ public class Search {
     private CubieCube[] preMoveCubes = new CubieCube[MAX_PRE_MOVES];
 
     // 值范围: [0, 17]
-    private int[] preMoves = new int[MAX_PRE_MOVES];
+    private byte[] preMoves = new byte[MAX_PRE_MOVES];
 
     private int preMoveLen = 0;
 
@@ -344,7 +344,7 @@ public class Search {
         }
 
         lm = lm / 3 * 3;
-        for (int m = 0; m < 18; m++) {
+        for (byte m = 0; m < 18; m++) {
             if (m == lm || m == lm - 9 || m == lm + 9) {
                 m += 2;
                 continue;
@@ -407,8 +407,8 @@ public class Search {
         int edgei = CubieCube.getPermSymInv(p2edge, p2esym, false);
         int corni = CubieCube.getPermSymInv(p2corn, p2csym, true);
 
-        int lastMove = depth == 0 ? -1 : move[depth - 1];
-        int lastPre = preMoveLen == 0 ? -1 : preMoves[preMoveLen - 1];
+        byte lastMove = depth == 0 ? -1 : move[depth - 1];
+        byte lastPre = preMoveLen == 0 ? -1 : preMoves[preMoveLen - 1];
 
         int ret = 0;
         int p2switchMax = (preMoveLen == 0 ? 1 : 2) * (depth == 0 ? 1 : 2);
@@ -428,7 +428,7 @@ public class Search {
             }
             if ((p2switch & 1) == 0 && depth > 0) {
                 int m = Util.std2ud[lastMove / 3 * 3 + 1];
-                move[depth - 1] = Util.ud2std[m] * 2 - move[depth - 1];
+                move[depth - 1] = (byte) (Util.ud2std[m] * 2 - move[depth - 1]);
 
                 p2mid = CoordCube.MPermMove[p2mid][m];
                 p2corn = CoordCube.CPermMove[p2corn][CubieCube.SymMoveUD[p2csym][m]];
@@ -441,7 +441,7 @@ public class Search {
                 edgei = CubieCube.getPermSymInv(p2edge, p2esym, false);
             } else if (preMoveLen > 0) {
                 int m = Util.std2ud[lastPre / 3 * 3 + 1];
-                preMoves[preMoveLen - 1] = Util.ud2std[m] * 2 - preMoves[preMoveLen - 1];
+                preMoves[preMoveLen - 1] = (byte) (Util.ud2std[m] * 2 - preMoves[preMoveLen - 1]);
 
                 p2mid = CubieCube.MPermInv[CoordCube.MPermMove[CubieCube.MPermInv[p2mid]][m]];
                 p2corn = CoordCube.CPermMove[corni >> 4][CubieCube.SymMoveUD[corni & 0xf][m]];
@@ -520,12 +520,12 @@ public class Search {
 
         int skipMoves = CubieCube.getSkipMoves(ssym);
 
-        for (int axis = 0; axis < 18; axis += 3) {
+        for (byte axis = 0; axis < 18; axis += 3) {
             if (axis == lm || axis == lm - 9) {
                 continue;
             }
-            for (int power = 0; power < 3; power++) {
-                int m = axis + power;
+            for (byte power = 0; power < 3; power++) {
+                byte m = (byte) (axis + power);
 
                 if (isRec && m != move[depth - maxl] || skipMoves != 0 && (skipMoves & 1 << m) != 0) {
                     continue;
@@ -600,12 +600,12 @@ public class Search {
 
         int skipMoves = CubieCube.getSkipMoves(ssym);
 
-        for (int axis = 0; axis < 18; axis += 3) {
+        for (byte axis = 0; axis < 18; axis += 3) {
             if (axis == lm || axis == lm - 9) {
                 continue;
             }
-            for (int power = 0; power < 3; power++) {
-                int m = axis + power;
+            for (byte power = 0; power < 3; power++) {
+                byte m = (byte) (axis + power);
 
                 if (isRec && m != move[length - maxl] || skipMoves != 0 && (skipMoves & 1 << m) != 0) {
                     continue;
@@ -720,7 +720,7 @@ public class Search {
         private int urfIdx = 0;
 
         // 值范围: [0, 17]
-        private int[] moves = new int[31];
+        private byte[] moves = new byte[31];
 
         public Solution() {
         }
@@ -735,7 +735,7 @@ public class Search {
             this.depth = depth;
         }
 
-        public void appendSolMove(int curMove) {
+        public void appendSolMove(byte curMove) {
             if (length == 0) {
                 moves[length++] = curMove;
                 return;
@@ -747,7 +747,7 @@ public class Search {
                 if (pow == 3) {
                     length--;
                 } else {
-                    moves[length - 1] = axisCur * 3 + pow;
+                    moves[length - 1] = (byte) (axisCur * 3 + pow);
                 }
                 return;
             }
@@ -757,7 +757,7 @@ public class Search {
                     length--;
                     moves[length - 1] = moves[length];
                 } else {
-                    moves[length - 2] = axisCur * 3 + pow;
+                    moves[length - 2] = (byte) (axisCur * 3 + pow);
                 }
                 return;
             }
